+++ NameSpaceDemo/app.js
@@ -1,0 +1,6 @@
+/// <reference path="utility-functions.ts" />
+var result = Utility.MaxBooksAllowed(15);
+console.log(result);
+var util = Utility.Fees;
+var fee = util.CalculateLateFee(10);
+console.log(fee);
