+++ NameSpaceDemo/index.html
@@ -1,0 +1,6 @@
+<html>
+    <head></head>
+    <body>
+        <script src="bundle.js"></script>       
+    </body>
+</html>
