+++ NameSpaceDemo/utility-functions.js
@@ -1,0 +1,17 @@
+var Utility;
+(function (Utility) {
+    var Fees;
+    (function (Fees) {
+        function CalculateLateFee(daysLate) {
+            return daysLate * 0.25;
+        }
+        Fees.CalculateLateFee = CalculateLateFee;
+    })(Fees = Utility.Fees || (Utility.Fees = {}));
+    function MaxBooksAllowed(age) {
+        return (age < 12) ? 3 : 10;
+    }
+    Utility.MaxBooksAllowed = MaxBooksAllowed;
+    function privateFunc() {
+        console.log('This is private');
+    }
+})(Utility || (Utility = {}));
