+++ src/app.ts
@@ -1,246 +1,16 @@
+import { UniversityLibrarian, RefBook, Shelf } from "./classes";
+import { Book, Author, Librarian, Magazine } from "./interfaces";
+import { Category } from "./enums";
+import { PersonBook } from "./types";
+import { getAllBooks, purge } from "./functions";
+
+
 const showHello = (divName: string, name: string) => {
   const elt = document.getElementById(divName);
   elt.innerText = `Hello from ${name}`;
 }
 
 showHello('greeting', 'TypeScript');
-
-// 1
-
-enum Category {
-  JavaScript,
-  CSS,
-  HTML,
-  TypeScript,
-  Angular
-}
-
-interface DamageLogger {
-  (reason: string): void;
-}
-
-interface Book {
-  id: number;
-  title: string;
-  author: string;
-  available: boolean;
-  category: Category;
-  pages?: number;
-  markDamaged?: DamageLogger;
-};
-
-
-
-const getAllBooks = () => {
-  const books: readonly Book[] = <const>[ 
-      {
-        id: 1,
-        title: 'Refactoring JavaScript',
-        author: 'Evan Burchard',
-        available: true,
-        category: Category.JavaScript
-      },
-      {
-        id: 2,
-        title: 'JavaScript Testing',
-        author: 'Liang Yuxian Eugene',
-        available: false,
-        category: Category.JavaScript
-      },
-      {
-        id: 3,
-        title: 'CSS Secrets',
-        author: 'Lea Verou',
-        available: true,
-        category: Category.CSS
-      },
-      {
-        id: 4,
-        title: 'Mastering JavaScript Object-Oriented Programming',
-        author: 'Andrea Chiarelli',
-        available: true,
-        category: Category.JavaScript
-      }    
-  ];
-
-  return books;
-}
-
-const logFirstAvailable = (books: readonly any[] = getAllBooks()): void => {
-  const numberOfBooks: number = books.length;
-  let firstAvailableBookTitle: string = '';
-
-  for (const currentBook of books) {
-    if (currentBook.available) {
-      firstAvailableBookTitle = currentBook.title;
-      break;
-    }
-  }
-
-  console.log(`Total Books: ${numberOfBooks}`);
-  console.log(`First Available Book: ${firstAvailableBookTitle}`);
-}
-
-logFirstAvailable(getAllBooks());
-
-const getBookTitlesByCategory = (category: Category = Category.JavaScript): Array<string> => {
-  console.log(`Getting books in category: ${Category[category]}`);
-
-  const allBooks = getAllBooks();
-  const filteredTitles: string[] = [];
-
-  for (let currentBook of allBooks) {
-    if (currentBook.category === category) {
-      filteredTitles.push(currentBook.title);
-    }
-  }
-
-  return filteredTitles;
-}
-
-const logBookTitles = (titles: string[]): void => {
-  for (const title of titles) {
-    console.log(title);
-  }
-}
-
-//logBookTitles(getBookTitlesByCategory(Category.JavaScript));
-
-const getBookAuthorByIndex = (index: number): [string, string] => {
-  const books = getAllBooks();
-  const { title, author } = books[index];
-  return [title, author];
-}
-
-/*const calcTotalPages = (): BigInt => {
-  const data = <const>[
-    { lib: 'libName1', books: 1_000_000_000, avgPagesPerBook: 250 },
-    { lib: 'libName2', books: 5_000_000_000, avgPagesPerBook: 300 },
-    { lib: 'libName3', books: 3_000_000_000, avgPagesPerBook: 280 }
-  ];
-
-  let result = data.reduce((acc: bigint, obj) => {
-    return acc + BigInt(obj.books) * BigInt(obj.avgPagesPerBook);
-  }, 0n);
-
-  return result;
-}*/
-
-// Task 03
-
-getBookTitlesByCategory(Category.JavaScript).forEach((val, id) => console.log(++id + ' - ' + val));
-
-type BookOrUndefined = Book | undefined;
-
-const getBookByID = (id: number): BookOrUndefined => {
-  return getAllBooks().find((item) => item.id === id)
-};
-
-console.log(getBookByID(1));
-
-const createCustomerID = (name: string, id: number): string => {
-  return `${name}${id}`;
-}
-
-const myID = createCustomerID('Ann', 10);
-console.log(myID);
-
-let IdGenerator = (name: string, id: number) => `${name}${id}`;
-
-IdGenerator = createCustomerID;
-
-console.log(IdGenerator('Ann', 10));
-
-const createCustomer = (name: string, age?: number, city?: string): void => {
-  console.log(`Creating customer ${name}`);
-
-  if (age) {
-    console.log(`Age: ${age}`);
-  }
-
-  if (city) {
-    console.log(`City: ${city}`);
-  }
-};
-
-createCustomer('Ann');
-createCustomer('Boris', 6);
-createCustomer('Clara', 12, 'Atlanta');
-
-console.log(getBookTitlesByCategory());
-
-const сheckoutBooks = (customer: string, ...bookIDs: number[]): string[] => {
-  console.log(`Checking out books for ${customer}`);
-
-  let booksCheckedOut: string[] = [];
-
-  for (const id of bookIDs) {
-    const book = getBookByID(id);
-    if (book && book.available) {
-      booksCheckedOut.push(book.title);
-    }
-  };
-
-  return booksCheckedOut;
-};
-
-const myBooks: string[] = сheckoutBooks('Ann', 1, 3, 4);
-console.log(myBooks);
-
-function getTitles(author: string): string[];
-function getTitles(available: boolean): string[];
-function getTitles(id: number, available: boolean): string[];
-
-function getTitles(...args: any[]): string[] {
-
-  const books: Readonly<any> = getAllBooks();
-
-  if (args.length === 0) {
-    return [];
-  }
-  else if (args.length === 1) {
-    const [arg] = args;
-
-    if (typeof arg === 'string') {
-      return books.filter(book => book.author === arg).map(book => book.title);
-    }
-    else if (typeof arg === 'boolean') {
-      return books.filter(book => book.available === arg).map(book => book.title);
-    }
-  }
-  else if (args.length === 2) {
-    const [id, available] = args; 
-
-    if (typeof id === 'number' && available === 'boolean') {
-      return books.filter(book => book.id === id && book.available === available).map(book => book.title);
-    }
-  }
-};
-
-const checkedOutBooks = getTitles(false);
-console.log(checkedOutBooks);
-
-function assertStringValue(val: any): asserts val is string {
-  if (typeof val !== 'string') {
-    throw new Error('value should have been a string.');
-  }
-};
-
-const bookTitleTransform = (title: any) => {
-  assertStringValue(title);
-
-  return [...title].reverse().join('');
-};
-
-console.log(bookTitleTransform(getAllBooks()[0].title));
-
-//console.log(bookTitleTransform(3));
-
-// 4
-
-const printBook = (book: Book): void => {
-  console.log(`${book.title} by ${book.author}`);
-};
 
 const myBook: Book = {
   id: 5,
@@ -252,28 +22,8 @@ const myBook: Book = {
   markDamaged: (reason: string) => console.log(`Damaged: ${reason}`),
 };
 
-printBook(myBook);
-myBook.markDamaged('missing back cover');
-
-// 4.02
-
-const logDamage = (reason: string) => console.log(`Damaged: ${reason}`);
-logDamage('missng back cover');
 
 // 4.03
-interface Person {
-  name: string;
-  email: string;
-};
-
-interface Author extends Person {
-  numBooksPublished: number;
-};
-
-interface Librarian extends Person {
-  department: string;
-  assistCustomer: (custName: string) => void;
-}
 
 const favoriteAuthor: Author = {
   email: 'ann@gmail.com',
@@ -300,43 +50,7 @@ console.log(offer.paper?.magazine);
 
 // 4.05
 
-type BookProperties = keyof Book;
-
-const getBookProp = (book: Book, prop: BookProperties): any => {
-  if (typeof book[prop] === 'function') {
-    return (book[prop] as Function).name;
-  }
-  return book[prop];
-}
-
-console.log(getBookProp(getAllBooks()[0], 'title'));        // Refactoring JavaScript
-console.log(getBookProp(getAllBooks()[0], 'markDamaged'));  // undefined
-// console.log(getBookProp(getAllBooks()[0], 'isbn'));      // error
-
-
 // 05.01
-abstract class ReferenceItem {
-  private _publisher: string;
-  static department: string = 'Research Dep';
-
-  constructor(public title: string, protected year: number) {  
-  }
-
-  printItem(): void {
-    console.log(`${this.title} was published in ${this.year}.`);   
-    console.log(`Department: ${ReferenceItem.department}`);   
-  }
-
-  get publisher(): string {
-    return this._publisher.toUpperCase();
-  }
-
-  set publisher(newPublisher: string) {
-    this._publisher = newPublisher;
-  }
-
-  abstract printCitation(): void;
-}
 
 /*const ref: ReferenceItem = new ReferenceItem('Updated Facts and Figures', 2016);
 ref.printItem();
@@ -346,22 +60,7 @@ console.log(ref.publisher); */
 
 // 5.02
 
-class Encyclopedia extends ReferenceItem {
-  constructor(newTitle: string, newYear: number, public edition: number) {
-    super(newTitle, newYear);
-  }
-
-  printItem(): void {
-    super.printItem();
-    console.log(`Edition: ${this.edition} (${this.year})`);
-  } 
-
-  printCitation(): void {
-    console.log(`${this.title} - ${this.year}`);
-  }
-}
-
-const refBook = new Encyclopedia('Hello Typescript', 2020, 1);
+const refBook = new RefBook('Hello Typescript', 2020, 1);
 refBook.printItem();
 
 // 5.03
@@ -370,32 +69,56 @@ console.log(refBook);
 
 // 5.04
 
-class UniversityLibrarian implements Librarian {
-
-  name: string;
-  email: string;
-  department: string;
-
-  assistCustomer(custName: string): void {
-    console.log(`${this.name} is assisting ${custName}`);
-  }
-}
-
 const favoriteLibrarian: Librarian = new UniversityLibrarian();
 favoriteLibrarian.name = 'Anna';
 favoriteLibrarian.assistCustomer('Boris');
 
 // 5.05
-type PersonBook = Person & Book;
 
 const personBook: PersonBook = {
-  name: 'Anna',
-  email: 'anna@example.com',
-  author: 'Boris',
-  available: true,
-  category: Category.HTML,
-  id: 1,
-  title: 'Introduction to HTML',
+    name: 'Anna',
+    email: 'anna@example.com',
+    author: 'Boris',
+    available: true,
+    category: Category.HTML,
+    id: 1,
+    title: 'Introduction to HTML',
 };
 
+// task 06.05
+import('./classes').then(module => {
+  const reader = new module.Reader();
+  reader.name = 'Anna';
+  reader.take(getAllBooks()[0]);
+  console.log(reader);
+});
 
+
+// task 07.01
+const inventory: Array<Book> = [
+  { id: 10, title: 'The C Programming Language', author: 'K & R', available: true, category: Category.Software },
+  { id: 11, title: 'Code Complete', author: 'Steve McConnell', available: true, category: Category.Software },
+  { id: 12, title: '8-Bit Graphics with Cobol', author: 'A. B.', available: true, category: Category.Software },
+  { id: 13, title: 'Cool autoexec.bat Scripts!', author: 'C. D.', available: true, category: Category.Software },
+];
+/*const result1 = purge<Book>(inventory);
+console.log(result1);
+
+const result2 = purge([1, 2, 3, 4]);
+console.log(result2);*/
+
+const bookShelf = new Shelf<Book>();
+inventory.forEach(book => bookShelf.add(book));
+const firstBook = bookShelf.getFirst();
+console.log( firstBook.title);
+
+const magazines: Array<Magazine> = [
+  { title: 'Programming Language Monthly', publisher: 'Code Mags' },
+  { title: 'Literary Fiction Quarterly', publisher: 'College Press' },
+  { title: 'Five Points', publisher: 'GSU' },
+];
+
+const magazineShelf = new Shelf<Magazine>();
+magazines.forEach(magazine => magazineShelf.add(magazine));
+const firstMagazine = magazineShelf.getFirst();
+console.log(firstMagazine.title);
