+++ src/classes/index.ts
@@ -1,0 +1,5 @@
+export { default as RefBook } from './encyclopedia';
+export * from './reference-item';
+export * from './university-librarian';
+export * from './reader';
+export { default as Shelf } from './shelf';
