+++ src/classes/reader.ts
@@ -1,0 +1,10 @@
+import { Book } from '../interfaces';
+
+export const Reader = class {
+    name: string;
+    books: Book[] = [];
+
+    take(book: Book): void {
+        this.books.push(book);
+    }
+};
