+++ src/enums.ts
@@ -1,0 +1,10 @@
+enum Category {
+    JavaScript,
+    CSS,
+    HTML,
+    TypeScript,
+    Angular,
+    Software,
+};
+
+export { Category };
